+++ components/note.tsx
@@ -451,12 +451,14 @@ export function Note({
           <DraggableRoot
             items={note.checklistItems ?? []}
             onItemsChange={(newItems) => {
-              handleReorderChecklistItems(note.id, newItems);
+              if (canEdit) {
+                handleReorderChecklistItems(note.id, newItems);
+              }
             }}
           >
             <DraggableContainer className="space-y-1">
               {note.checklistItems?.map((item) => (
-                <DraggableItem key={item.id} id={item.id}>
+                <DraggableItem key={item.id} id={item.id} disabled={!canEdit}>
                   <ChecklistItemComponent
                     item={item}
                     onToggle={handleToggleChecklistItem}
